% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aweSOM-package.R
\docType{package}
\name{aweSOM-package}
\alias{aweSOM-package}
\title{Interactive Self-Organizing Maps}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}
Interactive plots and interface for Kohonen self-organizing maps.
}
\references{
Kohonen T. (2001) \emph{Self-Organizing Maps}, 3rd edition,
  Springer Press, Berlin. <doi:10.1007/978-3-642-56927-2>
}
\seealso{
\code{\link{aweSOM}}, \code{\link{aweSOMplot}},
  \code{\link{somInit}}, \code{\link{somQuality}}.
}
