% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aweSOM.R
\name{aweSOM}
\alias{aweSOM}
\title{aweSOM interface}
\usage{
aweSOM()
}
\value{
No return value, used for side effects.
}
\description{
Launches the (offline) web-based interface for training and visualizing
self-organizing maps.
}
\details{
If the interface does not open automatically, open the printed link
  in a web browser.
  
  To open large files within the interface, use
  \code{options(shiny.maxRequestSize=2^30)} (or a suitably large file size)
  before lauching the interface.
}
\references{
Kohonen T. (2001) \emph{Self-Organizing Maps}, 3rd edition,
  Springer Press, Berlin. <doi:10.1007/978-3-642-56927-2>
}
