% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{cdt}
\alias{cdt}
\title{Complete disjunctive table}
\usage{
cdt(x)
}
\arguments{
\item{x}{\code{data.frame} on which the table is computed. All columns will
be treated as factors.}
}
\value{
A \code{matrix} of dummy variables, with \code{nrow(x)} rows and a
  number of columns equal to the sum of numbers of levels in all the
  variables of \code{x}.
}
\description{
Computes the complete disjunctive table of a set of factors, where each
factor (ie categorical variable) is encoded as a set of dummy variables, one
for each level (category).
}
