% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{somDist}
\alias{somDist}
\title{Distance measures on a SOM}
\usage{
somDist(som)
}
\arguments{
\item{som}{\code{kohonen} object, a SOM created by the \code{som} function.}
}
\value{
A \code{list} with distance measures: between cells on the grid,
  between prototypes in data space, and the neighborhood matrix on the grid.
}
\description{
Several distance measures between cells or prototypes of a trained SOM (in
grid space, in data space).
}
