\name{aweSOM}
\alias{aweSOM}

\title{Interactive interface and plots for self-organizing maps}
\description{Starts the aweSOM interface}
\usage{
  aweSOM()
}

\arguments{
  No arguments.
}

\value{This function starts the (shiny) interactive interface with the default 
  browser.
}

\author{Julien Boelaert \email{jubo.stats@gmail.com}}

\references{
  RStudio and Inc. (2013). shiny: Web Application Framework for R. R package
  version 0.7.0. \url{http://cran.r-project.org/package=shiny}
}
